requirements.txt
asyncio
websockets
python-escpos